Exclude NSObject protocol properties when generating accessors
class_copyPropertyList excludes superclass properties (which is what we want) but includes properties defined in conformed-to protocols. NSObject protocol at some point changed to defining some getters as properties, and we don't want to forward those to user defaults.

Noticed this when `po defaults` in LLDB returned an empty description.